Isolate forms code
