--- core/src/form/mod.rs
@@ -1,3 +1,0 @@
-mod modal;
-
-pub use modal::*;
