+++ core/src/forms/mod.rs
@@ -1,0 +1,11 @@
+//! Implementation of the forms system available in Minecraft.
+
+mod custom;
+mod menu;
+mod modal;
+mod elements;
+
+pub use modal::*;
+pub use menu::*;
+pub use elements::*;
+pub use custom::*;
